requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
python-multipart
pillow
torch
torchvision
python-jose[cryptography]
passlib[bcrypt]
alembic
pytest
httpx
dotenv
python-dotenvemail-validator
bcrypt==3.2.0
